requirements.txt
fastapi
uvicorn
pydantic
python-multipart
websockets
requests
google-generativeai